Izmene razne, izmena modela, repository, service...

+++ PIprojekat2019/src/main/java/projekat/repository/CenovnikRepository.java
@@ -1,10 +1,9 @@
 package projekat.repository;
 
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import projekat.model.Cenovnik;
 
-public interface CenovnikRepository extends JpaRepository<Cenovnik, Integer> {
+public interface CenovnikRepository extends JpaRepository<Cenovnik, Long> {
 
 }
