+++ PIprojekat2019/src/main/java/projekat/repository/FakturaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projekat.model.Faktura;
 
-public interface FakturaRepository extends JpaRepository<Faktura, Integer> {
+public interface FakturaRepository extends JpaRepository<Faktura, Long> {
 
 }
