+++ PIprojekat2019/src/main/java/projekat/repository/GrupaRobeUslugaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projekat.model.GrupaRobeUsluga;
 
-public interface GrupaRobeUslugaRepository extends JpaRepository<GrupaRobeUsluga, Integer> {
+public interface GrupaRobeUslugaRepository extends JpaRepository<GrupaRobeUsluga, Long> {
 
 }
