+++ PIprojekat2019/src/main/java/projekat/repository/JedinicaMereRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projekat.model.JedinicaMere;
 
-public interface JedinicaMereRepository extends JpaRepository<JedinicaMere, Integer> {
+public interface JedinicaMereRepository extends JpaRepository<JedinicaMere, Long> {
 
 }
