+++ PIprojekat2019/src/main/java/projekat/repository/NaseljenoMestoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projekat.model.NaseljenoMesto;
 
-public interface NaseljenoMestoRepository extends JpaRepository<NaseljenoMesto, Integer> {
+public interface NaseljenoMestoRepository extends JpaRepository<NaseljenoMesto, Long> {
 
 }
