+++ PIprojekat2019/src/main/java/projekat/repository/OtpremnicaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projekat.model.Otpremnica;
 
-public interface OtpremnicaRepository extends JpaRepository<Otpremnica, Integer> {
+public interface OtpremnicaRepository extends JpaRepository<Otpremnica, Long> {
 
 }
