+++ PIprojekat2019/src/main/java/projekat/repository/PDVKategorijaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projekat.model.PDVKategorija;
 
-public interface PDVKategorijaRepository extends JpaRepository<PDVKategorija, Integer> {
+public interface PDVKategorijaRepository extends JpaRepository<PDVKategorija, Long> {
 
 }
