+++ PIprojekat2019/src/main/java/projekat/repository/PDVStopaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projekat.model.PDVStopa;
 
-public interface PDVStopaRepository extends JpaRepository<PDVStopa, Integer> {
+public interface PDVStopaRepository extends JpaRepository<PDVStopa, Long> {
 
 }
