+++ PIprojekat2019/src/main/java/projekat/repository/PoslovnaGodinaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projekat.model.PoslovnaGodina;
 
-public interface PoslovnaGodinaRepository extends JpaRepository<PoslovnaGodina, Integer> {
+public interface PoslovnaGodinaRepository extends JpaRepository<PoslovnaGodina, Long> {
 
 }
