+++ PIprojekat2019/src/main/java/projekat/repository/PoslovniPartnerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projekat.model.PoslovniPartner;
 
-public interface PoslovniPartnerRepository extends JpaRepository<PoslovniPartner, Integer> {
+public interface PoslovniPartnerRepository extends JpaRepository<PoslovniPartner, Long> {
 
 }
