+++ PIprojekat2019/src/main/java/projekat/repository/PreduzeceRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projekat.model.Preduzece;
 
-public interface PreduzeceRepository extends JpaRepository<Preduzece, Integer> {
+public interface PreduzeceRepository extends JpaRepository<Preduzece, Long> {
 
 }
