+++ PIprojekat2019/src/main/java/projekat/repository/StavkaCenovnikaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projekat.model.StavkaCenovnika;
 
-public interface StavkaCenovnikaRepository extends JpaRepository<StavkaCenovnika, Integer> {
+public interface StavkaCenovnikaRepository extends JpaRepository<StavkaCenovnika, Long> {
 
 }
