+++ PIprojekat2019/src/main/java/projekat/repository/StavkaFaktureRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projekat.model.StavkaFakture;
 
-public interface StavkaFaktureRepository extends JpaRepository<StavkaFakture, Integer> {
+public interface StavkaFaktureRepository extends JpaRepository<StavkaFakture, Long> {
 
 }
