+++ PIprojekat2019/src/main/java/projekat/repository/StavkaOtpremniceRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projekat.model.StavkaOtpremnice;
 
-public interface StavkaOtpremniceRepository extends JpaRepository<StavkaOtpremnice, Integer> {
+public interface StavkaOtpremniceRepository extends JpaRepository<StavkaOtpremnice, Long> {
 
 }
